refactor: P2 openai forward/passthrough + gateway 闭包 TK companion
将 openai_gateway_forward 协议分流/Codex identity/body/headers、passthrough
failover/short-stream buffer，以及 gateway count_tokens/models/responses 闭包
抽到 *_tk_* companion；同步 protocol-routing SSOT 与 gateway-tk sentinel。
功能不变。no-web-impact
